[박하민] sprint6

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicChannelService.java
@@ -1,59 +1,73 @@
 package com.sprint.mission.discodeit.service.basic;
 
-import com.sprint.mission.discodeit.dto.ChannelDTO;
+import com.sprint.mission.discodeit.dto.ChannelDto;
 import com.sprint.mission.discodeit.dto.request.PrivateChannelCreateRequest;
 import com.sprint.mission.discodeit.dto.request.PublicChannelCreateRequest;
 import com.sprint.mission.discodeit.dto.request.PublicChannelUpdateRequest;
-import com.sprint.mission.discodeit.entity.*;
+import com.sprint.mission.discodeit.entity.Channel;
+import com.sprint.mission.discodeit.entity.ChannelType;
+import com.sprint.mission.discodeit.entity.ReadStatus;
+import com.sprint.mission.discodeit.mapper.ChannelMapper;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.MessageRepository;
 import com.sprint.mission.discodeit.repository.ReadStatusRepository;
+import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.ChannelService;
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.time.Instant;
-import java.util.*;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class BasicChannelService implements ChannelService {
 
   private final ChannelRepository channelRepository;
   private final ReadStatusRepository readStatusRepository;
   private final MessageRepository messageRepository;
+  private final UserRepository userRepository;
+  private final ChannelMapper channelMapper;
 
+  @Transactional
   @Override
-  public Channel create(PublicChannelCreateRequest channelCreateRequest) {
+  public ChannelDto create(PublicChannelCreateRequest channelCreateRequest) {
     String name = channelCreateRequest.name();
     String description = channelCreateRequest.description();
 
     Channel channel = new Channel(ChannelType.PUBLIC, name, description);
-    return channelRepository.save(channel);
+    channelRepository.save(channel);
+    return channelMapper.toDto(channel);
+  }
+
+  @Transactional
+  @Override
+  public ChannelDto create(PrivateChannelCreateRequest channelCreateRequest) {
+    Channel channel = new Channel(ChannelType.PRIVATE);
+    channelRepository.save(channel);
+
+    List<ReadStatus> readStatusList = userRepository.findAllById(
+            channelCreateRequest.participantIds())
+        .stream()
+        .map(user -> new ReadStatus(user, channel, channel.getCreatedAt()))
+        .toList();
+
+    readStatusRepository.saveAll(readStatusList);
+    return channelMapper.toDto(channel);
   }
 
   @Override
-  public Channel create(PrivateChannelCreateRequest channelCreateRequest) {
-    Channel channel = new Channel(ChannelType.PRIVATE, null, null);
-    Channel createdChannel = channelRepository.save(channel);
-
-    channelCreateRequest.participantIds().stream() // 채널의 참여자들을 스트림으로 변환
-        .map(userId -> new ReadStatus(userId, createdChannel.getId(),
-            channel.getCreatedAt())) // 각 참여자마다 readStatus 객체 생성
-        .forEach(readStatusRepository::save);
-
-    return createdChannel;
-  }
-
-  @Override
-  public ChannelDTO find(UUID channelId) {
+  public ChannelDto find(UUID channelId) {
     return channelRepository.findById(channelId)
-        .map(channel -> toChannelDTO(channel))
+        .map(channelMapper::toDto)
         .orElseThrow(() -> new NoSuchElementException(channelId + " 채널을 찾을 수 없습니다."));
   }
 
+  @Transactional
   @Override
-  public Channel update(UUID channelId, PublicChannelUpdateRequest channelUpdateRequest) {
+  public ChannelDto update(UUID channelId, PublicChannelUpdateRequest channelUpdateRequest) {
     String newName = channelUpdateRequest.newName();
     String newDescription = channelUpdateRequest.newDescription();
 
@@ -65,56 +79,30 @@ public class BasicChannelService implements ChannelService {
     }
 
     channel.update(newName, newDescription);
-    return channelRepository.save(channel);
+    return channelMapper.toDto(channel);
   }
 
+  @Transactional
   @Override
   public void delete(UUID channelId) {
-    Channel channel = channelRepository.findById(channelId)
-        .orElseThrow(() -> new NoSuchElementException(channelId + " 채널을 찾을 수 없습니다."));
+    if (!channelRepository.existsById(channelId)) {
+      throw new NoSuchElementException(channelId + " 채널을 찾을 수 없습니다.");
+    }
 
-    messageRepository.deleteAllByChannelId(channel.getId());
-    readStatusRepository.deleteAllByChannelId(channel.getId());
-
+    messageRepository.deleteAllByChannelId(channelId);
+    readStatusRepository.deleteAllByChannelId(channelId);
     channelRepository.deleteById(channelId);
   }
 
   @Override
-  public List<ChannelDTO> findAllByUserId(UUID userId) {
-    List<UUID> mySubscribedChannelIds = readStatusRepository.findAllByUserId(userId).stream()
-        .map(ReadStatus::getChannelId)
-        .toList();
-
-    return channelRepository.findAll().stream()
-        .filter(channel ->
-            channel.getType().equals(ChannelType.PUBLIC)
-                || mySubscribedChannelIds.contains(channel.getId())
-        )
-        .map(this::toChannelDTO)
-        .toList();
-  }
-
-  private ChannelDTO toChannelDTO(Channel channel) {
-        /*
-        [ ] 해당 채널의 가장 최근 메시지의 시간 정보를 포함합니다.
-        [ ] PRIVATE 채널인 경우 참여한 User의 id 정보를 포함합니다.
-        */
-    Instant lastMessageAt = messageRepository.findAllByChannelId(channel.getId())
+  public List<ChannelDto> findAllByUserId(UUID userId) {
+    List<UUID> mySubscribedChannelIds = readStatusRepository.findAllByUserIdWithFetchJoin(userId)
         .stream()
-        .sorted(Comparator.comparing(Message::getCreatedAt).reversed())
-        .map(Message::getCreatedAt)
-        .limit(1)
-        .findFirst()
-        .orElse(Instant.MIN);
+        .map(readStatus -> readStatus.getChannel().getId())
+        .toList();
 
-    List<UUID> participantIds = new ArrayList<>();
-    if (channel.getType().equals(ChannelType.PRIVATE)) {
-      readStatusRepository.findAllByChannelId(channel.getId())
-          .stream()
-          .map(ReadStatus::getUserId)
-          .forEach(participantIds::add);
-    }
-
-    return ChannelDTO.fromEntity(channel, lastMessageAt, participantIds);
+    return channelRepository.findAllPublicOrSubscribed(mySubscribedChannelIds).stream()
+        .map(channelMapper::toDto)
+        .toList();
   }
 }
